Extract GPUI conversation export helpers (Fixes #152)
