+++ src/ui_gpui/views/chat_view/mod.rs
@@ -14,6 +14,8 @@ mod command;
 mod ime;
 mod render;
 mod render_bars;
+mod render_bars_export;
+
 mod render_sidebar;
 mod state;
